Moved content container interfaces into their own files
To make these interfaces easier to find.
