+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/Anchor.kt
@@ -20,40 +20,9 @@
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
 import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
-import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
-import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
 import org.apache.maven.doxia.sink.Sink
-import org.apache.maven.doxia.sink.SinkEventAttributes
 
 class Anchor(override val sink: Sink) :
     DoxiaContent(),
     TextContainer,
     FigureGraphicsContainer
-
-interface AnchorContainer {
-    val sink: Sink
-
-    @JvmDefault
-    fun anchor(
-        name: String,
-        id: String = "",
-        cssClass: String = "",
-        style: Style? = null,
-        lang: String = "",
-        title: String = "",
-        init: Anchor.() -> Unit
-    ) {
-        sink.anchor(
-            name,
-            attributesOf(
-                SinkEventAttributes.ID to id,
-                SinkEventAttributes.CLASS to cssClass,
-                SinkEventAttributes.STYLE to style?.value,
-                SinkEventAttributes.LANG to lang,
-                SinkEventAttributes.TITLE to title
-            )
-        )
-        init(Anchor(sink))
-        sink.anchor_()
-    }
-}
