+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/DateContainer.kt
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,7 +19,15 @@
  */
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
-import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
 import org.apache.maven.doxia.sink.Sink
 
-class Body(override val sink: Sink) : DoxiaContent(), BlockContentContainer
+interface DateContainer {
+    val sink: Sink
+
+    @JvmDefault
+    fun date(init: Date.() -> Unit) {
+        sink.date()
+        init(Date(sink))
+        sink.date_()
+    }
+}
