+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/HorizontalRuleContainer.kt
