+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/LinkContainer.kt
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,30 +19,27 @@
  */
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
-import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
 import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
 import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
-import com.github.gantsign.maven.doxia.sink.kotlin.style.VAlign
 import org.apache.maven.doxia.sink.Sink
 import org.apache.maven.doxia.sink.SinkEventAttributes
 
-class Section(override val sink: Sink, private val level: Int) :
-    DoxiaContent(),
-    BlockContentContainer {
+interface LinkContainer {
+    val sink: Sink
 
-    fun title(
-        vAlign: VAlign? = null,
+    @JvmDefault
+    fun link(
+        name: String,
         id: String = "",
         cssClass: String = "",
         style: Style? = null,
         lang: String = "",
         title: String = "",
-        init: SectionTitle.() -> Unit
+        init: Link.() -> Unit
     ) {
-        sink.sectionTitle(
-            level,
+        sink.link(
+            name,
             attributesOf(
-                SinkEventAttributes.VALIGN to vAlign?.value,
                 SinkEventAttributes.ID to id,
                 SinkEventAttributes.CLASS to cssClass,
                 SinkEventAttributes.STYLE to style?.value,
@@ -50,7 +47,7 @@ class Section(override val sink: Sink, private val level: Int) :
                 SinkEventAttributes.TITLE to title
             )
         )
-        init(SectionTitle(sink))
-        sink.sectionTitle_(level)
+        init(Link(sink))
+        sink.link_()
     }
 }
