+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/NumberedListItem.kt
@@ -20,35 +20,6 @@
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
 import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
-import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
-import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
 import org.apache.maven.doxia.sink.Sink
-import org.apache.maven.doxia.sink.SinkEventAttributes
 
 class NumberedListItem(override val sink: Sink) : DoxiaContent(), BlockContentContainer
-
-interface NumberedListItemContainer {
-    val sink: Sink
-
-    @JvmDefault
-    fun listItem(
-        id: String = "",
-        cssClass: String = "",
-        style: Style? = null,
-        lang: String = "",
-        title: String = "",
-        init: NumberedListItem.() -> Unit
-    ) {
-        sink.numberedListItem(
-            attributesOf(
-                SinkEventAttributes.ID to id,
-                SinkEventAttributes.CLASS to cssClass,
-                SinkEventAttributes.STYLE to style?.value,
-                SinkEventAttributes.LANG to lang,
-                SinkEventAttributes.TITLE to title
-            )
-        )
-        init(NumberedListItem(sink))
-        sink.numberedListItem_()
-    }
-}
