+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/Paragraph.kt
@@ -20,10 +20,7 @@
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
 import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
-import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
-import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
 import org.apache.maven.doxia.sink.Sink
-import org.apache.maven.doxia.sink.SinkEventAttributes
 
 class Paragraph(override val sink: Sink) :
     DoxiaContent(),
@@ -31,31 +28,3 @@ class Paragraph(override val sink: Sink) :
     FigureGraphicsContainer,
     LinkContainer,
     AnchorContainer
-
-interface ParagraphContainer {
-    val sink: Sink
-
-    @JvmDefault
-    fun paragraph(
-        align: String = "",
-        id: String = "",
-        cssClass: String = "",
-        style: Style? = null,
-        lang: String = "",
-        title: String = "",
-        init: Paragraph.() -> Unit
-    ) {
-        sink.paragraph(
-            attributesOf(
-                SinkEventAttributes.ALIGN to align,
-                SinkEventAttributes.ID to id,
-                SinkEventAttributes.CLASS to cssClass,
-                SinkEventAttributes.STYLE to style?.value,
-                SinkEventAttributes.LANG to lang,
-                SinkEventAttributes.TITLE to title
-            )
-        )
-        init(Paragraph(sink))
-        sink.paragraph_()
-    }
-}
