+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/TableCaption.kt
@@ -20,37 +20,6 @@
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
 import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
-import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
-import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
 import org.apache.maven.doxia.sink.Sink
-import org.apache.maven.doxia.sink.SinkEventAttributes
 
 class TableCaption(override val sink: Sink) : DoxiaContent(), TextContainer
-
-interface TableCaptionContainer {
-    val sink: Sink
-
-    @JvmDefault
-    fun tableCaption(
-        align: String = "",
-        id: String = "",
-        cssClass: String = "",
-        style: Style? = null,
-        lang: String = "",
-        title: String = "",
-        init: TableCaption.() -> Unit
-    ) {
-        sink.tableCaption(
-            attributesOf(
-                SinkEventAttributes.ALIGN to align,
-                SinkEventAttributes.ID to id,
-                SinkEventAttributes.CLASS to cssClass,
-                SinkEventAttributes.STYLE to style?.value,
-                SinkEventAttributes.LANG to lang,
-                SinkEventAttributes.TITLE to title
-            )
-        )
-        init(TableCaption(sink))
-        sink.tableCaption_()
-    }
-}
