+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/TableCell.kt
@@ -20,59 +20,6 @@
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
 import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
-import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
-import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
 import org.apache.maven.doxia.sink.Sink
-import org.apache.maven.doxia.sink.SinkEventAttributes
 
 class TableCell(override val sink: Sink) : DoxiaContent(), BlockContentContainer
-
-interface TableCellContainer {
-    val sink: Sink
-
-    @JvmDefault
-    fun tableCell(
-        abbvr: String = "",
-        align: String = "",
-        axis: String = "",
-        bgColor: String = "",
-        colSpan: Int? = null,
-        headers: String = "",
-        height: String = "",
-        noWrap: String = "",
-        rowSpan: Int? = null,
-        scope: String = "",
-        vAlign: String = "",
-        width: String = "",
-        id: String = "",
-        cssClass: String = "",
-        style: Style? = null,
-        lang: String = "",
-        title: String = "",
-        init: TableCell.() -> Unit
-    ) {
-        sink.tableCell(
-            attributesOf(
-                SinkEventAttributes.ABBRV to abbvr,
-                SinkEventAttributes.ALIGN to align,
-                SinkEventAttributes.AXIS to axis,
-                SinkEventAttributes.BGCOLOR to bgColor,
-                SinkEventAttributes.COLSPAN to colSpan,
-                SinkEventAttributes.HEADERS to headers,
-                SinkEventAttributes.HEIGHT to height,
-                SinkEventAttributes.NOWRAP to noWrap,
-                SinkEventAttributes.ROWSPAN to rowSpan,
-                SinkEventAttributes.SCOPE to scope,
-                SinkEventAttributes.VALIGN to vAlign,
-                SinkEventAttributes.WIDTH to width,
-                SinkEventAttributes.ID to id,
-                SinkEventAttributes.CLASS to cssClass,
-                SinkEventAttributes.STYLE to style?.value,
-                SinkEventAttributes.LANG to lang,
-                SinkEventAttributes.TITLE to title
-            )
-        )
-        init(TableCell(sink))
-        sink.tableCell_()
-    }
-}
