+++ src/main/kotlin/com/github/gantsign/maven/doxia/sink/kotlin/content/TableContainer.kt
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,30 +19,43 @@
  */
 package com.github.gantsign.maven.doxia.sink.kotlin.content
 
-import com.github.gantsign.maven.doxia.sink.kotlin.DoxiaContent
 import com.github.gantsign.maven.doxia.sink.kotlin.internal.attributesOf
 import com.github.gantsign.maven.doxia.sink.kotlin.style.Style
-import com.github.gantsign.maven.doxia.sink.kotlin.style.VAlign
 import org.apache.maven.doxia.sink.Sink
 import org.apache.maven.doxia.sink.SinkEventAttributes
 
-class Section(override val sink: Sink, private val level: Int) :
-    DoxiaContent(),
-    BlockContentContainer {
+interface TableContainer {
+    val sink: Sink
 
-    fun title(
-        vAlign: VAlign? = null,
+    @JvmDefault
+    fun table(
+        align: String = "",
+        bgColor: String = "",
+        border: String = "",
+        cellPadding: String = "",
+        cellSpacing: String = "",
+        frame: String = "",
+        rules: String = "",
+        summary: String = "",
+        width: String = "",
         id: String = "",
         cssClass: String = "",
         style: Style? = null,
         lang: String = "",
         title: String = "",
-        init: SectionTitle.() -> Unit
+        init: Table.() -> Unit
     ) {
-        sink.sectionTitle(
-            level,
+        sink.table(
             attributesOf(
-                SinkEventAttributes.VALIGN to vAlign?.value,
+                SinkEventAttributes.ALIGN to align,
+                SinkEventAttributes.BGCOLOR to bgColor,
+                SinkEventAttributes.BORDER to border,
+                SinkEventAttributes.CELLPADDING to cellPadding,
+                SinkEventAttributes.CELLSPACING to cellSpacing,
+                SinkEventAttributes.FRAME to frame,
+                SinkEventAttributes.RULES to rules,
+                SinkEventAttributes.SUMMARY to summary,
+                SinkEventAttributes.WIDTH to width,
                 SinkEventAttributes.ID to id,
                 SinkEventAttributes.CLASS to cssClass,
                 SinkEventAttributes.STYLE to style?.value,
@@ -50,7 +63,7 @@ class Section(override val sink: Sink, private val level: Int) :
                 SinkEventAttributes.TITLE to title
             )
         )
-        init(SectionTitle(sink))
-        sink.sectionTitle_(level)
+        init(Table(sink))
+        sink.table_()
     }
 }
